Made sessions controller support turbo and devise 4.9
Devise 4.9 adds support for turbo, mainly by changing the response statuses in the sessions controller from 302 to 303 for success and 200 to 422 for errors.

    https://github.com/heartcombo/devise/wiki/How-To:-Upgrade-to-Devise-4.9.0-%5BHotwire-Turbo-integration%5D

I've made this work in the same way, so that it works with turbo.

However, there is another issue caused by the way that devise passwordless re-renders the new template after signin. Turbo doesn't support re-rendering after form-submission, if you try it shows this error in the browser console:

    Error: Form responses must redirect to another location

To resolve that, I've changed the sessions controller to redirect to new after signing in rather than re-rendering.