Bug fix for password not being required when used in tandem with the `database_authenticatable` and `validatable` modules.
This fix properly ensures that the `#password_required?` and `#password` methods are not redefined by `Devise::Models::MagicLinkAuthenticatable` if they already exist in the model. The check for the existing methods via `instance_methods.include?` must be done in the context of the class the module is being included into and not in the module itself.

See #13 for full context.