Extract send_magic_link out to a method
I've extracted the code to send the magic link out to it's own method, which allows it to be overridden in other apps using this gem if they want to perform additional actions when the email is sent successfully.

This can already be done using the `send_magic_link` method on the "user" model, but sometimes the logic you want to add doesn't make sense in the model layer.